refactor: adds warmUpTemplatesCache script
This script runs during console-api build time and prefetches all Akash templates into data folder.
The current code will just run this cache as it was created in runtime.
