+++ apps/api/src/services/external/githubService.ts
@@ -4,11 +4,8 @@ import minutesToSeconds from "date-fns/minutesToSeconds";
 
 import { cacheKeys, cacheResponse } from "@src/caching/helpers";
 import { Auditor, ProviderAttributesSchema } from "@src/types/provider";
-import { env } from "@src/utils/env";
 
-export function getOctokit() {
-  const githubPAT = env.GITHUB_PAT;
-
+export function getOctokit(githubPAT: string) {
   if (!githubPAT) {
     throw new Error("GITHUB_PAT is missing");
   }
