+++ packages/docker/docker-compose.build.yml
@@ -7,6 +7,7 @@ services:
       target: production-nginx
       args:
         WORKSPACE: apps/api
+        GITHUB_PAT: ${GITHUB_PAT}
 
   indexer:
     image: ${INDEXER_REPO:-console-indexer}:${INDEXER_TAG:-latest}
